Fix probs not sum to 1

+++ src/qibo/core/states.py
@@ -92,7 +92,7 @@ class VectorState(AbstractState):
     def probabilities(self, qubits=None, measurement_gate=None):
         unmeasured_qubits = tuple(i for i in range(self.nqubits)
                                   if i not in qubits)
-        state = K.reshape(K.square(K.abs(K.cast(self.tensor))), self.nqubits * (2,))
+        state = K.reshape(K.square(K.abs(K.cast(self.tensor, dtype="complex128"))), self.nqubits * (2,))
         return K.sum(state, axis=unmeasured_qubits)
 
     def measure(self, gate, nshots, registers=None):
